requirements.txt
Flask
google-generativeai
scikit-learn
numpy
python-dotenv